+++ src/test/java/org/pgcodekeeper/core/it/parser/ch/ChParserTest.java
@@ -52,7 +52,7 @@ class ChParserTest {
         "ch_view, 10",
         "ch_policy, 0",
         "ch_table, 3",
-        "ch_select, 180",
+        "ch_select, 179",
         "ch_privileges, 0",
         "ch_user, 0",
         "ch_role, 0",